requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.3
djangorestframework==3.14.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pygame==2.4.0
pytest==7.3.1
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
urllib3==2.0.3
